rename package to eliott
