+++ eliott/utils.py
@@ -2,10 +2,10 @@ import json
 import os
 from pathlib import Path
 
-CACHE_DIR = Path(os.path.expanduser("~")) / ".cache_mp3"
+CACHE_DIR = Path(os.path.expanduser("~")) / ".cache_eliott"
 CACHE_DIR.mkdir(exist_ok=True)
 CACHE_FILE = CACHE_DIR / "cache.json"
-CONFIG_PATH = Path(os.path.expanduser("~")) / ".mp3organizer_config"
+CONFIG_PATH = Path(os.path.expanduser("~")) / ".eliott_config"
 
 
 def load_cache():
@@ -34,4 +34,26 @@ def save_config(cfg: dict) -> None:
 
 def sanitize_filename(name: str) -> str:
     return ''.join(c for c in name if c.isalnum() or c in (' ', '-', '_')).title()
+
+
+def unique_path(path: Path) -> Path:
+    """Return a unique file path by appending an incrementing suffix.
+
+    If ``path`` does not exist it is returned unchanged. Otherwise ``_1``,
+    ``_2`` ... are appended before the file extension until a path that does
+    not exist is found.
+    """
+    if not path.exists():
+        return path
+
+    base = path.stem
+    suffix = path.suffix
+    parent = path.parent
+    index = 1
+    while True:
+        candidate = parent / f"{base}_{index}{suffix}"
+        if not candidate.exists():
+            return candidate
+        index += 1
+
