+++ src/main.py
@@ -1,4 +1,4 @@
-from mp3organizer.gui import main
+from eliott.gui import main
 
 if __name__ == '__main__':
     main()
