+++ tests/test_core.py
@@ -3,7 +3,7 @@ import sys
 from pathlib import Path
 
 sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
-from mp3organizer import core
+from eliott import core
 
 class DummySpotify:
     def me(self):
